test(LottoTest): separate responsibility about profit rate and move to statistics test class
